cabana: move stream open widgets into streamselector (#38715)
