+++ openpilot/tools/cabana/streams/socketcanstream.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <QComboBox>
-
 #include "tools/cabana/streams/livestream.h"
 
 struct SocketCanStreamConfig {
@@ -24,18 +22,4 @@ protected:
 
   SocketCanStreamConfig config = {};
   int sock_fd = -1;
-};
-
-class OpenSocketCanWidget : public AbstractOpenStreamWidget {
-  Q_OBJECT
-
-public:
-  OpenSocketCanWidget(QWidget *parent = nullptr);
-  AbstractStream *open() override;
-
-private:
-  void refreshDevices();
-
-  QComboBox *device_edit;
-  SocketCanStreamConfig config = {};
 };
